fix: refactor task types by moving them to a dedicated types file and updating imports across components
