+++ src/types/date.ts
@@ -1,0 +1,2 @@
+export type Fecha =
+  `${number}-${number}-${number} ${number}:${number}:${number}`;
